+++ example/src/TranscribeData.tsx
@@ -6,7 +6,7 @@ import RNFS from 'react-native-fs'
 import Sound from 'react-native-sound'
 import { initWhisper, libVersion } from '../../src'
 import type { WhisperContext } from '../../src'
-import { WavFileWriter } from '../../src/realtime-transcription/WavFileWriter'
+import { WavFileWriter } from '../../src/utils/WavFileWriter'
 import { Button } from './Button'
 import contextOpts from './context-opts'
 import { createDir, fileDir } from './util'
