+++ src/utils/WavFileWriter.ts
@@ -1,4 +1,4 @@
-import { base64ToUint8Array, uint8ArrayToBase64 } from './utils'
+import { base64ToUint8Array, uint8ArrayToBase64 } from './common'
 
 export interface WavFileConfig {
   sampleRate: number
